Extract renderBootstrappedToasts function to toasts.ts [DEV-55]
It's easy to imagine this code being reused in other apps.
